+++ packages/api-client/src/auth/PreKey.ts
@@ -18,6 +18,8 @@
  */
 
 export interface PreKey {
+  /** The PreKey ID */
   id: number;
+  /** The PreKey data, base64 encoded */
   key: string;
 }