+++ packages/api-client/src/http/HttpClient.ts
@@ -22,17 +22,28 @@ import {PriorityQueue} from '@wireapp/priority-queue';
 import axios, {AxiosError, AxiosRequestConfig, AxiosResponse} from 'axios';
 import {EventEmitter} from 'events';
 import logdown from 'logdown';
-import {AccessTokenData, AccessTokenStore, AuthAPI} from '../auth/';
-import {BackendErrorMapper, ConnectionState, ContentType, NetworkError, StatusCode} from '../http/';
+import {StatusCodes as HTTP_STATUS} from 'http-status-codes';
+
+import {
+  AccessTokenData,
+  AccessTokenStore,
+  AuthAPI,
+  InvalidTokenError,
+  MissingCookieError,
+  TokenExpiredError,
+} from '../auth/';
+import {BackendErrorMapper, ConnectionState, ContentType, NetworkError} from '../http/';
 import {ObfuscationUtil} from '../obfuscation/';
 import {sendRequestWithCookie} from '../shims/node/cookie';
 
 enum TOPIC {
   ON_CONNECTION_STATE_CHANGE = 'HttpClient.TOPIC.ON_CONNECTION_STATE_CHANGE',
+  ON_INVALID_TOKEN = 'HttpClient.TOPIC.ON_INVALID_TOKEN',
 }
 
 export interface HttpClient {
   on(event: TOPIC.ON_CONNECTION_STATE_CHANGE, listener: (state: ConnectionState) => void): this;
+  on(event: TOPIC.ON_INVALID_TOKEN, listener: (error: InvalidTokenError | MissingCookieError) => void): this;
 }
 
 export class HttpClient extends EventEmitter {
@@ -126,16 +137,29 @@ export class HttpClient extends EventEmitter {
       if (response) {
         const {data: errorData, status: errorStatus} = response;
         const isBackendError = errorData?.code && errorData?.label && errorData?.message;
-
         if (isBackendError) {
           error = BackendErrorMapper.map(errorData);
-        } else {
-          const isUnauthorized = errorStatus === StatusCode.UNAUTHORIZED;
-          const hasAccessToken = !!this.accessTokenStore?.accessToken;
-          if (isUnauthorized && hasAccessToken && firstTry) {
-            await this.refreshAccessToken();
-            return this._sendRequest<T>(config, tokenAsParam, false);
-          }
+        }
+
+        const isExpiredTokenError = error instanceof TokenExpiredError;
+        const isUnauthorized = errorStatus === HTTP_STATUS.UNAUTHORIZED;
+        const hasAccessToken = !!this.accessTokenStore?.accessToken;
+
+        if ((isExpiredTokenError || isUnauthorized) && hasAccessToken && firstTry) {
+          this.logger.warn(
+            `Access token refresh triggered (isExpiredTokenError: ${isExpiredTokenError}, isUnauthorized: ${isUnauthorized}) for "${config.method}" request to "${config.url}".`,
+          );
+          await this.refreshAccessToken();
+          return this._sendRequest<T>(config, tokenAsParam, false);
+        }
+
+        if (error instanceof InvalidTokenError || error instanceof MissingCookieError) {
+          // On invalid cookie the application is supposed to logout.
+          this.logger.warn(
+            `Cannot renew access token for "${config.method}" request to "${config.url}" because cookie/token is invalid: ${error.message}`,
+            error,
+          );
+          this.emit(HttpClient.TOPIC.ON_INVALID_TOKEN, error);
         }
       }
 
@@ -190,7 +214,7 @@ export class HttpClient extends EventEmitter {
       ...config.headers,
       'Content-Type': ContentType.APPLICATION_JSON,
     };
-    return this.sendRequest<T>(config, isSynchronousRequest);
+    return this.sendRequest<T>(config, false, isSynchronousRequest);
   }
 
   public sendXML<T>(config: AxiosRequestConfig, isSynchronousRequest: boolean = false): Promise<AxiosResponse<T>> {
@@ -198,7 +222,7 @@ export class HttpClient extends EventEmitter {
       ...config.headers,
       'Content-Type': ContentType.APPLICATION_XML,
     };
-    return this.sendRequest<T>(config, isSynchronousRequest);
+    return this.sendRequest<T>(config, false, isSynchronousRequest);
   }
 
   public sendProtocolBuffer<T>(
@@ -209,6 +233,6 @@ export class HttpClient extends EventEmitter {
       ...config.headers,
       'Content-Type': ContentType.APPLICATION_PROTOBUF,
     };
-    return this.sendRequest<T>(config, isSynchronousRequest);
+    return this.sendRequest<T>(config, false, isSynchronousRequest);
   }
 }