+++ packages/bot-handler-uptime/src/UptimeHandler.ts
@@ -20,10 +20,10 @@
 import * as moment from 'moment';
 import momentDurationFormatSetup from 'moment-duration-format';
 
-import {PayloadBundle, PayloadBundleSource, PayloadBundleType} from '@wireapp/core/dist/conversation/';
-import {QuotableMessage} from '@wireapp/core/dist/conversation/message/OtrMessage';
+import {PayloadBundle, PayloadBundleSource, PayloadBundleType} from '@wireapp/core/src/main/conversation/';
+import {QuotableMessage} from '@wireapp/core/src/main/conversation/message/OtrMessage';
 import {MessageHandler} from '@wireapp/bot-api';
-import {TextContent} from '@wireapp/core/dist/conversation/content';
+import {TextContent} from '@wireapp/core/src/main/conversation/content';
 
 momentDurationFormatSetup(moment as any);
 
@@ -43,7 +43,7 @@ export class UptimeHandler extends MessageHandler {
     switch (payload.type) {
       case PayloadBundleType.TEXT: {
         const content = payload.content as TextContent;
-        if (content.text === '/uptime') {
+        if (content.text.trim() === '/uptime') {
           const upTime = UptimeHandler.getUptime();
           await this.sendReply(payload.conversation, payload as QuotableMessage, `Running since: ${upTime}`);
         }